Add basic logic for add package action to get package data
